refactor: changed header to a separated component
